+++ src/headless/utils/init.js
@@ -9,7 +9,7 @@ import { Connection } from '@converse/headless/shared/connection/index.js';
 import { Model } from '@converse/skeletor/src/model.js';
 import { Strophe } from 'strophe.js/src/strophe';
 import { createStore, initStorage } from '@converse/headless/utils/storage.js';
-import { getLoginCredentialsFromBrowser } from '@converse/headless/utils/core.js';
+import { isValidJID } from './core.js';
 
 
 export function initPlugins (_converse) {
@@ -87,17 +87,6 @@ export async function initSessionStorage (_converse) {
     };
 }
 
-export async function initScramStorage (_converse) {
-    _converse.storage = {
-        ..._converse.storage,
-        'scramStorage': Storage.localForage.createInstance({
-            'name': 'converse-scram',
-            'description': 'SCRAM storage driver',
-            'driver': Storage.localForage.INDEXEDDB
-        })
-    };
-}
-
 function initPersistentStorage (_converse, store_name) {
     if (_converse.api.settings.get('persistent_store') === 'sessionStorage') {
         return;
@@ -169,6 +158,7 @@ function saveJIDtoSession (_converse, jid) {
  */
 export async function setUserJID (jid) {
     await initSession(_converse, jid);
+
     /**
      * Triggered whenever the user's JID has been updated
      * @event _converse#setUserJID
@@ -306,18 +296,24 @@ export async function attemptNonPreboundSession (credentials, automatic) {
         // automatically setting up a new session (``auto_login``).
         // So we can't do the check (!automatic || _converse.api.settings.get("auto_login")) here.
         if (credentials) {
-            connect(credentials);
+            return connect(credentials);
         } else if (api.settings.get("credentials_url")) {
             // We give credentials_url preference, because
             // _converse.connection.pass might be an expired token.
-            connect(await getLoginCredentials());
+            return connect(await getLoginCredentials());
         } else if (_converse.jid && (api.settings.get("password") || _converse.connection.pass)) {
-            connect();
-        } else if (!_converse.isTestEnv() && 'credentials' in navigator) {
-            connect(await getLoginCredentialsFromBrowser());
-        } else {
-            !_converse.isTestEnv() && log.warn("attemptNonPreboundSession: Couldn't find credentials to log in with");
+            return connect();
         }
+
+        if (api.settings.get('save_scram_keys')) {
+            const credentials = await getLoginCredentialsFromSCRAMKeys();
+            if (credentials) return connect(credentials);
+        }
+
+        if (!_converse.isTestEnv() && 'credentials' in navigator) {
+            return connect(await getLoginCredentialsFromBrowser());
+        }
+        !_converse.isTestEnv() && log.warn("attemptNonPreboundSession: Couldn't find credentials to log in with");
     } else if (
         [_converse.ANONYMOUS, _converse.EXTERNAL].includes(api.settings.get("authentication")) &&
         (!automatic || api.settings.get("auto_login"))
@@ -338,7 +334,52 @@ export function getConnectionServiceURL () {
 }
 
 
-function connect (credentials) {
+export async function getLoginCredentialsFromBrowser () {
+    try {
+        const creds = await navigator.credentials.get({'password': true});
+        if (creds && creds.type == 'password' && isValidJID(creds.id)) {
+            await setUserJID(creds.id);
+            return {'jid': creds.id, 'password': creds.password};
+        }
+    } catch (e) {
+        log.error(e);
+    }
+}
+
+
+export async function getLoginCredentialsFromSCRAMKeys () {
+    const jid = localStorage.getItem('conversejs-session-jid');
+    if (!jid) {
+        return null;
+    }
+    await setUserJID(jid);
+
+    const login_info = await savedLoginInfo(jid);
+    const scram_keys = login_info.get('scram_keys');
+    return scram_keys ? { jid , 'password': scram_keys } : null;
+}
+
+
+/**
+ * Fetch the stored SCRAM keys for the given JID, if available.
+ *
+ * The user's plaintext password is not stored, nor any material from which
+ * the user's plaintext password could be recovered.
+ *
+ * @param { String } JID - The XMPP address for which to fetch the SCRAM keys
+ * @returns { Promise } A promise which resolves once we've fetched the previously
+ *  used login keys.
+ */
+export async function savedLoginInfo (jid) {
+    const id = `converse.scram-keys-${Strophe.getBareJidFromJid(jid)}`;
+    const login_info = new Model({ id });
+    initStorage(login_info, id, 'persistent');
+    await new Promise(f => login_info.fetch({'success': f, 'error': f}));
+    return login_info;
+}
+
+
+async function connect (credentials) {
     const { api } = _converse;
     if ([_converse.ANONYMOUS, _converse.EXTERNAL].includes(api.settings.get("authentication"))) {
         if (!_converse.jid) {
@@ -369,26 +410,14 @@ function connect (credentials) {
 
         let callback;
 
-        if (api.settings.get("save_scram_keys") && !password.ck) {
-            // Don't save the SCRAM data if we already logged in with SCRAM
-            const login_info = await _converse.api.savedLoginInfo();
+        // Save the SCRAM data if we're not already logged in with SCRAM
+        if (_converse.jid && api.settings.get("save_scram_keys") && !password?.ck) {
+            // Store scram keys in scram storage
+            const login_info = await savedLoginInfo(_converse.jid);
 
-            callback = async (status) => {
-                // Store scram keys in scram storage
-                if (!_converse?.storage?.scramStorage) {
-                    await initScramStorage(_converse);
-                }
-
-                const newScramKeys = _converse.connection.scramKeys;
-                if (newScramKeys) {
-                    try {
-                        const new_users_info = login_info.users ?? { };
-                        new_users_info[_converse.connection.authzid] = newScramKeys;
-                        login_info.save({'users': new_users_info });
-                    } catch (e) { // Could not find local storage }
-                        log.error("No storage method found: ", e);
-                    }
-                }
+            callback = (status) => {
+                const { scram_keys } = _converse.connection;
+                if (scram_keys) login_info.save({ scram_keys });
                 _converse.connection.onConnectStatusChanged(status);
             };
         }