+++ CHANGES.md
@@ -5,7 +5,10 @@
 - #326: Add the ability to reset your password
 - #2816: Chat highlight behaves odd
 - #2925: File upload is not always enabled
+- #3001: Add option to save SCRAM details and to use them to stay logged in upon reload
 - Add a "Add to Contacts" button in MUC occupant modals
+
+- New config option [save_scram_keys](https://conversejs.org/docs/html/configuration.html#save-scram-keys)
 
 ## 10.0.0 (2022-10-30)